move components and css
