--- app/css/components/entrybox.css
@@ -1,19 +1,0 @@
-.entrybox {
-  width: 100%;
-  margin-bottom: 72px;
-}
-
-.header {
-  composes: heading from '../common/typography';
-
-  text-align: center;
-}
-
-.input {
-  composes: heading from '../common/typography';
-  composes: input from '../common/input';
-
-  text-align: center;
-  outline: none;
-}
-
