--- app/css/components/main-section.css
@@ -1,20 +1,0 @@
-@value global-radius from '../settings.css';
-
-.main-section {
-  composes: card from '../common/card';
-  width: 50%;
-  border-radius: global-radius;
-}
-
-  .header {
-    composes: title from '../common/card';
-    composes: subheading from '../common/typography';
-    background: #ed193a;
-    color: #fff;
-  }
-
-  .list {
-    list-style: none;
-    padding: 16px;
-  }
-
