+++ app/topics/Vote.jsx
@@ -1,9 +1,9 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 import classNames from 'classnames/bind';
-import EntryBox from 'components/EntryBox';
-import MainSection from 'components/MainSection';
-import Scoreboard from 'components/Scoreboard';
+import EntryBox from './components/EntryBox';
+import MainSection from './components/MainSection';
+import Scoreboard from './components/Scoreboard';
 import { createTopic, typing, incrementCount,
   decrementCount, destroyTopic, fetchTopics } from './actions';
 import styles from './style.css';
