+++ app/topics/components/EntryBox.jsx
@@ -1,7 +1,7 @@
 import React, { PropTypes } from 'react';
-import TopicTextInput from 'components/TopicTextInput';
+import TopicTextInput from './TopicTextInput';
 import classNames from 'classnames/bind';
-import styles from 'css/components/entrybox';
+import styles from './entrybox';
 
 const cx = classNames.bind(styles);
