+++ app/topics/components/entrybox.css
@@ -1,0 +1,19 @@
+.entrybox {
+  width: 100%;
+  margin-bottom: 72px;
+}
+
+.header {
+  composes: heading from 'css/common/typography';
+
+  text-align: center;
+}
+
+.input {
+  composes: heading from 'css/common/typography';
+  composes: input from 'css/common/input';
+
+  text-align: center;
+  outline: none;
+}
+
