+++ app/users/LoginOrRegister.jsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import classNames from 'classnames/bind';
 import { connect } from 'react-redux';
 import { manualLogin, signUp, toggleLoginMode } from 'users/actions';
-import styles from 'css/components/login';
+import styles from './login';
 import hourGlassSvg from 'images/hourglass.svg';
 
 const cx = classNames.bind(styles);
